refactor(transactions): home TypeFormProps in props.ts, unify balance effect timer
