+++ features/transactions/entry/typeForms/props.ts
@@ -1,0 +1,10 @@
+// features/transactions/entry/typeForms/props.ts
+import type { DraftAction, DraftState } from '../draftReducer';
+
+export type TypeFormProps = {
+  draft: DraftState;
+  dispatch: (a: DraftAction) => void;
+  accounts: string[];
+  payees: string[];
+  defaultCurrency: string;
+};
